requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==2.2.10
django-extensions==2.2.9
djangorestframework==3.11.0
flake8==3.8.2
importlib-metadata==1.6.0
mccabe==0.6.1
pathspec==0.8.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.5.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
zipp==3.1.0
